Explicitly cast elementIds to int in Graph/MutableGraph

+++ source/app/graph/graph.cpp
@@ -227,7 +227,7 @@ void Graph::reserveNodeId(NodeId nodeId)
     _nextNodeId = nodeId + 1;
 
     for(auto nodeArray : _nodeArrays)
-        nodeArray->resize(_nextNodeId);
+        nodeArray->resize(static_cast<int>(_nextNodeId));
 }
 
 void Graph::reserveEdgeId(EdgeId edgeId)
@@ -238,7 +238,7 @@ void Graph::reserveEdgeId(EdgeId edgeId)
     _nextEdgeId = edgeId + 1;
 
     for(auto edgeArray : _edgeArrays)
-        edgeArray->resize(_nextEdgeId);
+        edgeArray->resize(static_cast<int>(_nextEdgeId));
 }
 
 void Graph::clear()
